+++ src/details.js
@@ -1,5 +1,5 @@
 const Apify = require('apify'); // eslint-disable-line no-unused-vars
-const { parseCaption } = require('./helpers');
+const { parseCaption, mapNode, mapProp, edgesToText, dedupById, secondsToDate } = require('./helpers');
 const consts = require('./consts');
 
 const { PAGE_TYPES } = consts;
@@ -16,8 +16,8 @@ const formatIGTVVideo = (edge) => {
         type: 'Video',
         shortCode: node.shortcode,
         title: node.title,
-        caption: node?.edge_media_to_caption?.edges?.length ? node.edge_media_to_caption.edges[0].node.text : '',
-        commentsCount: node?.edge_media_to_comment?.count,
+        caption: edgesToText(node.edge_media_to_caption?.edges).join('\n') || '',
+        commentsCount: node.edge_media_to_comment?.count,
         commentsDisabled: node.comments_disabled,
         dimensionsHeight: node.dimensions.height,
         dimensionsWidth: node.dimensions.width,
@@ -33,19 +33,14 @@ const formatIGTVVideo = (edge) => {
  * @param {Array<Record<string, any>>} resources
  */
 const formatDisplayResources = (resources) => {
-    if (!resources) return [];
-    return resources.map((resource) => resource.node.display_url).filter((s) => s);
+    return (resources ?? []).map(mapNode).map(mapProp('display_url')).filter(Boolean);
 };
 
 /**
  *
  * @param {Record<string, any>} node
  */
-const sidecarImages = (node) => {
-    return node?.edge_sidecar_to_children?.edges?.length
-        ? formatDisplayResources(node.edge_sidecar_to_children.edges)
-        : [];
-};
+const sidecarImages = (node) => formatDisplayResources(node.edge_sidecar_to_children?.edges);
 
 /**
  * Format Post Edge item into cleaner output
@@ -59,20 +54,16 @@ const formatSinglePost = (node) => {
             node.edge_media_to_parent_comment?.count,
             node.edge_media_preview_comment?.count,
             node.edge_media_to_hoisted_comment?.count,
-        ].reduce((out, count) => (out + (count ?? 0)), 0),
-        edges: [
+        ].reduce((out, count) => (count > out ? count : out), 0), // Math.max won't work here
+        edges: dedupById([
             node.edge_media_to_comment?.edges,
             node.edge_media_to_parent_comment?.edges,
             node.edge_media_preview_comment?.edges,
             node.edge_media_to_hoisted_comment?.edges,
-        ].reduce((out, edges) => {
-            return out.concat(edges ?? []);
-        }, []),
+        ].flat().map(mapNode)),
     };
     const likes = node.edge_liked_by || node.edge_media_preview_like;
-    const caption = node?.edge_media_to_caption?.edges?.length
-        ? node.edge_media_to_caption.edges.reduce((out, { node: { text } }) => ([...out, text]), []).join('\n')
-        : '';
+    const caption = edgesToText(node.edge_media_to_caption?.edges).join('\n');
     const { hashtags, mentions } = parseCaption(caption);
 
     return {
@@ -83,24 +74,25 @@ const formatSinglePost = (node) => {
         caption,
         hashtags,
         mentions,
-        url: `https://www.instagram.com/p/${node.shortcode}`,
-        commentsCount: comments?.count ?? 0,
-        latestComments: comments?.edges?.length ? comments.edges.map((edge) => ({
-            ownerUsername: edge?.node?.owner?.username ?? '',
-            text: edge.node.text,
-        })).reverse() : [],
+        url: node.shortcode ? `https://www.instagram.com/p/${node.shortcode}/` : null,
+        commentsCount: comments.count || comments.edges.length || 0,
+        firstComment: comments?.edges?.[0]?.text ?? '',
+        latestComments: comments.edges.map(({ id, owner, text, created_at }) => ({
+            id,
+            ownerUsername: owner?.username ?? '',
+            text,
+            timestamp: secondsToDate(created_at),
+        })),
         dimensionsHeight: node.dimensions.height,
         dimensionsWidth: node.dimensions.width,
         displayUrl: node.display_url,
         images: sidecarImages(node),
         videoUrl: node.video_url,
-        firstComment: comments?.edges?.[0]?.node?.text ?? '',
         alt: node.accessibility_caption,
         likesCount: likes?.count ?? null,
         videoViewCount: node.video_view_count,
-        timestamp: node.taken_at_timestamp
-            ? new Date(node.taken_at_timestamp * 1000).toISOString()
-            : null,
+        timestamp: secondsToDate(node.taken_at_timestamp),
+        childPosts: node.edge_sidecar_to_children?.edges?.map?.((child) => formatSinglePost(child.node)) ?? [],
         locationName: node.location?.name ?? null,
         locationId: node.location?.id ?? null,
         ownerFullName: node?.owner?.full_name ?? null,